setup linting and formatting
