+++ .eslintrc.json
@@ -4,24 +4,19 @@
     "es6": true,
     "node": true
   },
-  "extends": [
-    "eslint:recommended",
-    "plugin:react/recommended"
-  ],
+  "extends": ["eslint:recommended", "prettier"],
   "globals": {
     "Atomics": "readonly",
     "SharedArrayBuffer": "readonly"
   },
   "parserOptions": {
     "ecmaFeatures": {
-        "jsx": true
+      "jsx": true
     },
     "ecmaVersion": 2018,
     "sourceType": "module"
   },
-  "plugins": [
-    "react"
-  ],
+  "plugins": ["html", "import"],
   "rules": {
     "array-bracket-spacing": ["error", "always"],
     "curly": ["error", "multi-line"],
@@ -35,6 +30,8 @@
     "no-undef": "error",
     "object-curly-spacing": ["error", "always"],
     "quotes": ["error", "single"],
-    "semi": ["error", "always"]
+    "semi": ["error", "always"],
+    "no-unused-vars": "warn",
+    "no-console": "off"
   }
 }
