+++ css/styles.css
@@ -1,6 +1,6 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto:ital,wght@0,100..900;1,100..900&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100..900;1,100..900&display=swap");
-@import "tailwindcss";
+@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto:ital,wght@0,100..900;1,100..900&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100..900;1,100..900&display=swap');
+@import 'tailwindcss';
 
 @theme {
   --color-primary: #4caf50;
@@ -19,8 +19,8 @@
   --color-success-background: #e8f5e9;
   --color-failed: #c62828;
   --color-failed-background: #ffebee;
-  --font-Poppins: "Poppins", "sans-serif";
-  --font-Roboto: "Roboto", "sans-serif";
+  --font-Poppins: 'Poppins', 'sans-serif';
+  --font-Roboto: 'Roboto', 'sans-serif';
 }
 /* CUSTOM CLASSES */
 
@@ -29,7 +29,7 @@
 .primary-button {
   background-color: #4caf50;
   color: #ffffff;
-  font-family: "Poppins", "sans-serif";
+  font-family: 'Poppins', 'sans-serif';
   font-size: larger;
   font-weight: 500;
   letter-spacing: 0.05em;
@@ -47,7 +47,7 @@
 .secondary-button {
   border: 2px solid #2196f3;
   color: #2196f3;
-  font-family: "Poppins", "sans-serif";
+  font-family: 'Poppins', 'sans-serif';
   font-size: larger;
   font-weight: 500;
   letter-spacing: 0.05em;
@@ -63,7 +63,7 @@
 .alert-button {
   background-color: #ff5722;
   color: #ffffff;
-  font-family: "Poppins", "sans-serif";
+  font-family: 'Poppins', 'sans-serif';
   font-size: larger;
   font-weight: 500;
   letter-spacing: 0.05em;
