+++ package.json
@@ -1,37 +1,46 @@
 {
-	"name": "edu-gate",
-	"version": "1.0.0",
-	"description": "The frontend part of edugate",
-	"engines": {
-		"node": ">=18.17.1",
-		"npm": ">=10.8.2"
-	},
-	"main": "app.js",
-	"scripts": {
-		"build": "npx @tailwindcss/cli -i ./css/styles.css -o ./dist/output.css",
-		"tailwind:watch": "npx @tailwindcss/cli -i ./css/styles.css -o ./dist/output.css --watch",
-		"dev": "npm run build && node ./src/index.js && open ."
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/sulenchy/edu-gate.git"
-	},
-	"author": "",
-	"license": "ISC",
-	"bugs": {
-		"url": "https://github.com/sulenchy/edu-gate/issues"
-	},
-	"homepage": "https://github.com/sulenchy/edu-gate#readme",
-	"dependencies": {
-		"@tailwindcss/cli": "^4.0.1",
-		"edu-gate": "file:"
-	},
-	"directories": {
-		"doc": "docs"
-	},
-	"keywords": [],
-	"devDependencies": {
-		"live-server": "^1.2.0",
-		"tailwindcss": "^4.0.14"
-	}
+  "name": "edu-gate",
+  "version": "1.0.0",
+  "description": "The frontend part of edugate",
+  "engines": {
+    "node": ">=18.17.1",
+    "npm": ">=10.8.2"
+  },
+  "main": "app.js",
+  "scripts": {
+    "build": "npx @tailwindcss/cli -i ./css/styles.css -o ./dist/output.css",
+    "tailwind:watch": "npx @tailwindcss/cli -i ./css/styles.css -o ./dist/output.css --watch",
+    "dev": "npm run build && node ./src/index.mjs && open .",
+    "lint": "eslint .",
+    "format": "prettier --write ."
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sulenchy/edu-gate.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/sulenchy/edu-gate/issues"
+  },
+  "homepage": "https://github.com/sulenchy/edu-gate#readme",
+  "dependencies": {
+    "@tailwindcss/cli": "^4.0.1",
+    "edu-gate": "file:"
+  },
+  "directories": {
+    "doc": "docs"
+  },
+  "keywords": [],
+  "devDependencies": {
+    "@eslint/js": "^9.24.0",
+    "eslint": "^9.24.0",
+    "eslint-plugin-css-modules": "^2.12.0",
+    "eslint-plugin-html": "^8.1.2",
+    "eslint-plugin-import": "^2.31.0",
+    "globals": "^16.0.0",
+    "live-server": "^1.2.0",
+    "prettier": "^3.5.3",
+    "tailwindcss": "^4.0.14"
+  }
 }
