+++ tailwind.config.js
@@ -1,6 +1,7 @@
+/* global module */
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ["./**/*.html"],
+  content: ['./**/*.html'],
   theme: {
     extend: {},
   },
